refactor(gateway): GAR-477 — Q10.c extract build_discord_channels to bootstrap/discord.rs
Extracts build_discord_channels (165 LOC) + handle_discord_command (105 LOC)
from bootstrap/mod.rs into a new sibling module bootstrap/discord.rs.
Re-exported at the same public path — zero behaviour change.
bootstrap/mod.rs shrinks 2316 → 2044 lines (−272).

Follows the extract-and-re-export pattern from slice 10.a (config.rs, PR #89)
and slice 10.b (channels.rs, PR #470).

Plans/README bookkeeping: mark 0163 Done (PR #466), add 0164 (GAR-456 partial
fix PR #469), add 0164b (GAR-476 PR #470), add 0165 (this slice).

Linear: GAR-477
