+++ crates/garraia-gateway/src/bootstrap/mod.rs
@@ -23,6 +23,7 @@ use crate::state::SharedState;
 
 mod channels;
 mod config;
+mod discord;
 
 // Slice 10.a (GAR-440): path resolvers and API-key precedence chain extracted
 // to `bootstrap::config`. Re-exported at this level so external paths
@@ -33,6 +34,9 @@ pub(crate) use config::{default_vault_path, resolve_api_key};
 
 // Slice 10.b (GAR-476): channel registry orchestrator extracted to `bootstrap::channels`.
 pub use channels::build_channels;
+
+// Slice 10.c (GAR-477): Discord wiring + command handler extracted to `bootstrap::discord`.
+pub use discord::build_discord_channels;
 
 /// Build a fully-configured `AgentRuntime` from the application config.
 pub fn build_agent_runtime(config: &AppConfig) -> AgentRuntime {
@@ -1103,175 +1107,6 @@ pub async fn build_mcp_tools(config: &AppConfig) -> (McpManager, Vec<Box<dyn Too
     (manager, all_tools)
 }
 
-/// Build Discord channels from config. Must be called after state is
-/// wrapped in `Arc` so the message callback can capture a `SharedState`.
-pub fn build_discord_channels(
-    config: &AppConfig,
-    state: &SharedState,
-) -> Vec<Box<dyn garraia_channels::Channel>> {
-    let mut channels = Vec::new();
-
-    for (name, channel_config) in &config.channels {
-        if channel_config.channel_type != "discord" || channel_config.enabled == Some(false) {
-            continue;
-        }
-
-        // Inject secrets from env vars into the settings map.
-        let mut settings = channel_config.settings.clone();
-        if let Ok(token) = std::env::var("DISCORD_BOT_TOKEN") {
-            settings.insert("bot_token".to_string(), serde_json::json!(token));
-        }
-        if let Ok(app_id) = std::env::var("DISCORD_APP_ID")
-            && let Ok(id) = app_id.parse::<u64>()
-        {
-            settings.insert("application_id".to_string(), serde_json::json!(id));
-        }
-
-        let allowlist = Arc::new(Mutex::new(Allowlist::load_or_create(
-            &default_allowlist_path(),
-        )));
-        let pairing = Arc::new(Mutex::new(PairingManager::new(
-            std::time::Duration::from_secs(300),
-        )));
-
-        let state_for_cb = Arc::clone(state);
-        let allowlist_for_cb = Arc::clone(&allowlist);
-        let pairing_for_cb = Arc::clone(&pairing);
-
-        let on_message: garraia_channels::discord::DiscordOnMessageFn = Arc::new(
-            move |channel_id: String,
-                  user_id: String,
-                  user_name: String,
-                  text: String,
-                  delta_tx: Option<tokio::sync::mpsc::Sender<String>>| {
-                let state = Arc::clone(&state_for_cb);
-                let allowlist = Arc::clone(&allowlist_for_cb);
-                let pairing = Arc::clone(&pairing_for_cb);
-                Box::pin(async move {
-                    if let Some(cmd) = text.strip_prefix('/') {
-                        let cmd = cmd.split_whitespace().next().unwrap_or("");
-                        return handle_discord_command(
-                            cmd,
-                            &user_id,
-                            &user_name,
-                            &channel_id,
-                            &allowlist,
-                            &pairing,
-                            &state,
-                        );
-                    }
-
-                    {
-                        let mut list = allowlist.lock().unwrap();
-                        if list.needs_owner() {
-                            list.claim_owner(&user_id);
-                            info!("discord: auto-paired owner {} ({})", user_name, user_id);
-                            return Ok(format!(
-                                "Welcome, {}! You are now the owner of this GarraIA bot.\n\n\
-                                 Use /pair to generate a code for adding other users.\n\
-                                 Use /help for available commands.",
-                                user_name
-                            ));
-                        }
-
-                        if !list.is_allowed(&user_id) {
-                            let trimmed = text.trim();
-                            if trimmed.len() == 6 && trimmed.chars().all(|c| c.is_ascii_digit()) {
-                                let claimed = pairing.lock().unwrap().claim(trimmed, &user_id);
-                                if claimed.is_some() {
-                                    list.add(&user_id);
-                                    info!(
-                                        "discord: paired user {} ({}) via code",
-                                        user_name, user_id
-                                    );
-                                    return Ok(format!(
-                                        "Welcome, {}! You now have access to this bot.",
-                                        user_name
-                                    ));
-                                }
-                            }
-
-                            warn!(
-                                "discord: unauthorized user {} ({}) in channel {}",
-                                user_name, user_id, channel_id
-                            );
-                            return Err("__blocked__".to_string());
-                        }
-                    }
-
-                    let session_id = format!("discord-{channel_id}");
-
-                    let text = garraia_security::InputValidator::sanitize(&text);
-                    if garraia_security::InputValidator::check_prompt_injection(&text) {
-                        return Err(
-                            "input rejected: potential prompt injection detected".to_string()
-                        );
-                    }
-
-                    state
-                        .hydrate_session_history(&session_id, Some("discord"), Some(&user_id))
-                        .await;
-                    let history: Vec<ChatMessage> = state.session_history(&session_id);
-                    let continuity_key = state.continuity_key(Some(&user_id));
-
-                    let response = if let Some(delta_sender) = delta_tx {
-                        state
-                            .agents
-                            .process_message_streaming_with_context(
-                                &session_id,
-                                &text,
-                                &history,
-                                delta_sender,
-                                continuity_key.as_deref(),
-                                Some(&user_id),
-                                None,
-                            )
-                            .await
-                    } else {
-                        state
-                            .agents
-                            .process_message_with_context(
-                                &session_id,
-                                &text,
-                                &history,
-                                continuity_key.as_deref(),
-                                Some(&user_id),
-                            )
-                            .await
-                    }
-                    .map_err(|e| e.to_string())?;
-
-                    state
-                        .persist_turn(
-                            &session_id,
-                            Some("discord"),
-                            Some(&user_id),
-                            &text,
-                            &response,
-                        )
-                        .await;
-
-                    Ok(response)
-                })
-            },
-        );
-
-        match garraia_channels::discord::DiscordChannel::from_settings_with_callback(
-            &settings, on_message,
-        ) {
-            Ok(channel) => {
-                channels.push(Box::new(channel) as Box<dyn garraia_channels::Channel>);
-                info!("configured discord channel: {name}");
-            }
-            Err(e) => {
-                warn!("failed to configure discord channel {name}: {e}");
-            }
-        }
-    }
-
-    channels
-}
-
 /// Build a voice handler for Telegram that processes voice messages.
 /// Returns None if voice mode is not enabled (no STT/TTS clients).
 pub fn build_telegram_voice_handler(state: &SharedState) -> Option<OnVoiceFn> {
@@ -1700,113 +1535,6 @@ fn handle_command(
             Ok(format!("💥 Internal error: {msg}"))
         }
         Err(garraia_channels::CommandError::Blocked) => Err("__blocked__".to_string()),
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn handle_discord_command(
-    cmd: &str,
-    user_id: &str,
-    user_name: &str,
-    channel_id: &str,
-    allowlist: &Arc<Mutex<Allowlist>>,
-    pairing: &Arc<Mutex<PairingManager>>,
-    state: &SharedState,
-) -> std::result::Result<String, String> {
-    let list = allowlist.lock().unwrap();
-    let is_owner = list.is_owner(user_id);
-    let is_allowed = list.is_allowed(user_id);
-    drop(list);
-
-    match cmd {
-        "start" => {
-            if is_allowed {
-                Ok(
-                    "Welcome to GarraIA! Send me a message and I will respond.\n\n\
-                    Commands:\n\
-                    /help - show this help\n\
-                    /clear - reset conversation history\n\
-                    /pair - generate invite code (owner only)"
-                        .to_string(),
-                )
-            } else {
-                let mut list = allowlist.lock().unwrap();
-                if list.needs_owner() {
-                    list.claim_owner(user_id);
-                    info!("discord: auto-paired owner {} ({})", user_name, user_id);
-                    Ok(format!(
-                        "Welcome, {}! You are now the owner of this GarraIA bot.\n\n\
-                         Use /pair to generate a code for adding other users.",
-                        user_name
-                    ))
-                } else {
-                    Ok("This bot is private. Send the 6-digit pairing code you received to get access.".to_string())
-                }
-            }
-        }
-        "help" => {
-            if !is_allowed {
-                return Err("__blocked__".to_string());
-            }
-            let mut help = "GarraIA Commands:\n\
-                /help - show this help\n\
-                /clear - reset conversation history"
-                .to_string();
-            if is_owner {
-                help.push_str(
-                    "\n/pair - generate a 6-digit invite code\n/users - list allowed users",
-                );
-            }
-            Ok(help)
-        }
-        "clear" => {
-            if !is_allowed {
-                return Err("__blocked__".to_string());
-            }
-            let session_id = format!("discord-{channel_id}");
-            if let Some(mut session) = state.sessions.get_mut(&session_id) {
-                session.history.clear();
-            }
-            Ok("Conversation history cleared.".to_string())
-        }
-        "pair" => {
-            if !is_owner {
-                if !is_allowed {
-                    return Err("__blocked__".to_string());
-                }
-                return Ok("Only the bot owner can generate pairing codes.".to_string());
-            }
-            let code = pairing.lock().unwrap().generate("discord");
-            Ok(format!(
-                "Pairing code: {code}\n\n\
-                 Share this with the person you want to invite. \
-                 They should send this code to the bot within 5 minutes."
-            ))
-        }
-        "users" => {
-            if !is_owner {
-                if !is_allowed {
-                    return Err("__blocked__".to_string());
-                }
-                return Ok("Only the bot owner can list users.".to_string());
-            }
-            let list = allowlist.lock().unwrap();
-            let users = list.list_users();
-            let owner = list.owner().unwrap_or("none");
-            Ok(format!(
-                "Owner: {owner}\nAllowed users ({}):\n{}",
-                users.len(),
-                users.join("\n")
-            ))
-        }
-        _ => {
-            if !is_allowed {
-                return Err("__blocked__".to_string());
-            }
-            Ok(format!(
-                "Unknown command: /{cmd}\nUse /help for available commands."
-            ))
-        }
     }
 }
